Update Gradients.json to make a missing gradient
While working for a company I needed a most important gradient of black to white but it was not here. It looks so professional with the Black and White color selecting Black and White as the theme colors. Therefore and since in demand and because of its uniqueness I added this gradient