Update gradients.json
A great blend of a vibrant pink, a soft light blue, and a white cream color. I was actually using it for an NFT Marketplace "Coming Soon" page and thought that it would make a great, eye-catching background for the UI. With these colors. I thought a fitting name that is relatable and nostalgic to us all would be "Italian Ice" representing the traditional Watermelon, Blue Raspberry, and Lemon Flavors.